Add typescript-eslint and drop eslint to v8
The v9 flat config is not fun to use.
Expo's eslint config still does not support it which causes clashes with
the typescript-eslint config when using the v9 flat config.

So, I dropped back to eslint v8.

Also, expo's eslint config uses an old version of typescript-eslint
so I had to drop my dependencies to that version